Use NativePRNGNonBlocking for JRuby.
This changes the behavior on JRuby to first try to initialize a
`SecureRandom` instance using the `NativePRNGNonBlocking` algorithm
introduced in Java 8. This is based on `/dev/urandom`, making it roughly
equivalent to sysrandom's behavior in other Ruby implementations.

If `NativePRNGNonBlocking` is not available, it falls back on the
default `SecureRandom` configuration, whatever that may be.

Closes #10.